update: restruct proof process
