+++ cmd-tools/bin/relay.ts
@@ -1,0 +1,43 @@
+import Relay from "../lib/relay";
+import { queue } from "../lib/queue";
+import { config } from "../cfg";
+
+// get eth headers
+async function getHeaders(headers: any[], block: number) {
+  let header = await this.web3.eth.getBlock(block);
+  if (headers[0] === this.lastBlock) {
+    headers = headers.slice(1);
+    headers.push(header);
+  } else {
+    // restart process
+    this.relayService = false;
+    return;
+  }
+
+  await getHeaders(headers, block + 1);
+}
+
+(async function() {
+  // use the relay seed
+  config.sudo = config.relaySeed;
+
+  // set relay service
+  let relay = new Relay(config);
+  relay.relayService = true;
+  await relay.init();
+
+  // async process, fix this array's length to double times
+  // of blocksSafe.
+  //
+  // if the blocksSafe is 0, the value will be lucky 7.
+  let blocksLimit = relay.blocksSafe != 0 ?
+    relay.blocksSafe * 2 : 7;
+  let headers = new Array(blocksLimit);
+  headers.map(() => null);
+
+  // async without await
+  getHeaders(headers, relay.lastBlock.height + 1);
+
+  // relay process
+  queue.call(relay, 6);
+})();
