--- cmd-tools/lib/bootstrap.ts
@@ -1,6 +1,0 @@
-import API from "../../src/ethereum/API";
-
-export async function bootstrap(): Promise<void> {
-    const api = new API();
-    await api.setPolkadotJs();
-}
